requirements.txt
# Flask
flask
Flask-Assets
Flask-WTF
Flask-Login
flask-admin
flask_mail

# Database
Flask-SQLAlchemy
sqlalchemy-migrate
psycopg2

# documentation
eralchemy

# misc
python-dotenv

# Data Generation
factory_boy
Faker

readline


# Testing
nose
mamba
nosexcover
pylint

# Admin
flask-admin

# Payment
braintree

google-api-python-client
googlemaps

alembic
Flask-Script
Flask-Migrate
